Added cache entry linked list Added C++ unit tests for cache entry list Cache class is ready for integration into new server code

+++ SharpPacketLib/ICache.cs
@@ -3,7 +3,7 @@
     public interface ICache
     {
         Task<T?> GetAsync<T>(string key) where T : class;
-        Task<bool> SetAsync(string key, object value, double cacheSeconds);
+        Task<bool> SetAsync(string key, object value, int cacheSeconds);
         Task<bool> DelAsync(string key);
     }
 }
